Minor path bug fix
